[util] Move http-related util functions to `http.go` file
